abstracted out addRecipe
